CoreSignEditListener: Corrected order of `/sign wax` permission check
The permission check has to come after the override of the
honeycomb-on-sign check or else the native honeycomb behavior will be
ignored if the player does not have the `/sign wax` permission.

We won't interfere with the native behavior because it would be too
confounding and increase the chance of conflicts with other plugins.