Refactor sign interaction logic and fix sneaking behavior
This commit refactors the sign interaction logic in
`CoreSignEditListener.java` by simplifying the parameters of the
`overrideNativeBehavior()` method.  The `Player` is now directly
obtained from the event.

Moreover, this commit also adds a condition to block the native sign
interaction emulation when the player is sneaking because in the real
native implementation, the interaction would not take place.